refactor: move local storage logic to settings component
